client/requirements.txt
streamlit==1.32.0
requests==2.31.0
pandas==2.1.0
python-dotenv==1.0.0
black==23.9.1
ruff==0.0.292
openpyxl==3.1.2
plotly==5.18.0
